<SWAUTO-3464> Convert CallsMultiMap/IDataHandler interaction into Publisher-Subscriber pattern

+++ asam_cmp_capture_module/src/stream_fb.cpp
@@ -81,7 +81,7 @@ void StreamFb::updateStreamIdInternal()
     streamIdsList.erase(streamId);
     auto oldId = streamId;
 
-    Int newId = objPtr.getPropertyValue("StreamId");
+    uint8_t newId = static_cast<Int>(objPtr.getPropertyValue("StreamId"));
 
     if (streamIdManager->isValidId(newId))
     {
