+++ asam_cmp_common_lib/src/interface_common_fb.cpp
@@ -107,7 +107,7 @@ void InterfaceCommonFb::updatePayloadTypeInternal()
 void InterfaceCommonFb::removeStreamInternal(size_t nInd)
 {
     auto fb = functionBlocks.getItems().getItemAt(nInd);
-    Int streamId = fb.getPropertyValue("StreamId");
+    uint8_t streamId = static_cast<Int>(fb.getPropertyValue("StreamId"));
     functionBlocks.removeItem(fb);
     streamIdManager->removeId(streamId);
 }
